requirements.txt
appdirs==1.4.3
appnope==0.1.0
asgiref==3.4.1
asn1crypto==0.22.0
asttokens==2.0.5
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
black==22.3.0
cffi==1.15.0
click==8.0.4
coverage==4.5.1
cryptography==2.8
cycler==0.10.0
decorator==4.1.2
dj-database-url==0.4.2
Django==3.2.14
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-timedeltatemplatefilter==0.1.2
django-upgrade==1.0.0
ecdsa==0.13
enum34==1.1.6
et-xmlfile==1.0.1
executing==0.8.3
Fabric3==1.14.post1
flake8==4.0.1
future==0.16.0
idna==2.6
ipaddress==1.0.18
ipdb==0.13.9
ipython==7.16.3
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.17.2
kiwisolver==1.0.1
lml==0.0.1
matplotlib==3.0.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.5
odfpy==1.3.5
openpyxl==2.5.5
packaging==16.8
paramiko==2.7.1
parso==0.7.0
pathlib2==2.3.0
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.5.2
pure-eval==0.2.2
pyasn1==0.3.4
pycodestyle==2.8.0
pycparser==2.18
pydotplus==2.0.2
pyexcel==0.5.6
pyexcel-io==0.5.4
pyexcel-ods==0.5.2
pyexcel-xls==0.5.4
pyexcel-xlsx==0.5.4
pyflakes==2.4.0
Pygments==2.12.0
PyNaCl==1.1.2
pyparsing==2.2.0
python-dateutil==2.7.5
pytz==2017.2
scipy==1.5.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.3.5
sqlparse==0.3.0
stack-data==0.3.0
texttable==0.9.1
tokenize-rt==4.2.1
toml==0.10.2
tomli==1.2.3
traitlets==4.3.3
typing_extensions==4.1.1
uWSGI==2.0.15
wcwidth==0.2.5
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
xlwt-future==0.8.0
